requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
click==8.1.8
    # via flask
colorama==0.4.6
    # via click
flask==3.1.0
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
numpy==2.2.1
    # via pandas
pandas==2.2.3
    # via -r requirements.in
psycopg2==2.9.10
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
six==1.17.0
    # via python-dateutil
tzdata==2024.2
    # via pandas
werkzeug==3.1.3
    # via flask
